@@ -25,12 +25,12 @@
 		"@babel/register": "^7.5.5",
 		"@sentry/webpack-plugin": "^1.8.0",
 		"autoprefixer": "^9.6.1",
-		"babel-eslint": "^10.0.2",
+		"babel-eslint": "^10.0.3",
 		"babel-loader": "^8.0.6",
-		"babel-preset-minify": "^0.5.0",
+		"babel-preset-minify": "^0.5.1",
 		"cssnano": "^4.1.10",
-		"eslint": "^6.1.0",
-		"eslint-loader": "^2.2.1",
+		"eslint": "^6.3.0",
+		"eslint-loader": "^3.0.0",
 		"favicons": "^5.4.1",
 		"grunt": "^1.0.4",
 		"grunt-browser-sync": "^2.2.0",
@@ -41,7 +41,7 @@
 		"grunt-contrib-watch": "^1.1.0",
 		"grunt-mkdir": "^1.0.0",
 		"grunt-postcss": "^0.9.0",
-		"grunt-sass": "^3.0.2",
+		"grunt-sass": "^3.1.0",
 		"grunt-template": "^1.0.0",
 		"grunt-webpack": "^3.1.3",
 		"grunt-zopfli": "^0.3.3",
@@ -50,12 +50,12 @@
 		"npm-check": "^5.9.0",
 		"pixrem": "^5.0.0",
 		"svg-inline-loader": "^0.8.0",
-		"webpack": "^4.39.1",
-		"webpack-build-notifier": "^1.0.3",
+		"webpack": "^4.39.3",
+		"webpack-build-notifier": "^1.1.1",
 		"webpack-bundle-analyzer": "^3.4.1"
 	},
 	"dependencies": {
-		"@sentry/browser": "^5.6.1",
+		"@sentry/browser": "^5.6.3",
 		"autonumeric": "^4.5.4",
 		"backbone": "^1.4.0",
 		"filesize": "^4.1.2",
