@@ -1,10 +1,30 @@
 'use strict';
 
-const Backbone = require('backbone');
-const _ = require('underscore');
-const AppView = require('../views/app');
-const StoreHelper = require('../helpers/store');
-const DataHelper = require('../helpers/data');
+import Backbone from 'backbone';
+import _ from 'underscore';
+import AppView from '../views/app';
+import StoreHelper from '../helpers/store';
+import DataHelper from '../helpers/data';
+
+import BudgetView from '../views/budget';
+import DocumentSettingsAccountAddManualView from '../views/documentSettingsAccountAddManual';
+import DocumentSettingsAccountAddView from '../views/documentSettingsAccountAddManual';
+import DocumentSettingsAccountDetailsView from '../views/documentSettingsAccountDetails';
+import DocumentSettingsAccountView from '../views/documentSettingsAccount';
+import DocumentSettingsGeneralView from '../views/documentSettingsGeneral';
+import DocumentSettingsPermissionsView from '../views/documentSettingsPermissions';
+import DocumentSettingsPluginAddDetailsView from '../views/documentSettingsPluginAddDetails';
+import DocumentSettingsPluginAddSetupView from '../views/documentSettingsPluginAddSetup';
+import DocumentSettingsPluginAddView from '../views/documentSettingsPluginAdd';
+import DocumentSettingsPluginDetailsView from '../views/documentSettingsPluginDetails';
+import DocumentSettingsPluginView from '../views/documentSettingsPlugin';
+import DocumentSettingsView from '../views/documentSettings';
+import ErrorView from '../views/error';
+import PlaceHolderView from '../views/placeholder';
+import TransactionsView from '../views/transactions';
+import PluginModel from '../models/plugin';
+import PluginInstanceModel from '../models/pluginInstance';
+
 
 /**
  * @module routers/app
@@ -12,7 +32,7 @@ const DataHelper = require('../helpers/data');
  * @augments Backbone.Router
  * @author Sebastian Pekarek
  */
-module.exports = Backbone.Router.extend({
+const AppRouter = Backbone.Router.extend({
     routes: {
         'admin': 'adminRedirect',
         'admin/': 'adminDashboard',
@@ -80,7 +100,6 @@ module.exports = Backbone.Router.extend({
 
 
     async documentRedirect () {
-        const DataHelper = require('../helpers/data');
         const documents = DataHelper.getDocuments();
         if (!documents.length) {
             await document.wait();
@@ -110,19 +129,16 @@ module.exports = Backbone.Router.extend({
     },
 
     budget (documentId) {
-        const BudgetView = require('../views/budget');
         this.view.renderView(new BudgetView({
             documentId
         }));
     },
     transactions (documentId) {
-        const TransactionsView = require('../views/transactions');
         this.view.renderView(new TransactionsView({
             documentId
         }));
     },
     reports (documentId) {
-        const PlaceHolderView = require('../views/placeholder');
         this.view.renderView(new PlaceHolderView({
             icon: 'paper-plane',
             text: `Document ${documentId.substr(0, 8)}: Reports`
@@ -130,87 +146,70 @@ module.exports = Backbone.Router.extend({
     },
 
     documentSettings (documentId) {
-        const DocumentSettingsView = require('../views/documentSettings');
         this.view.renderView(new DocumentSettingsView({
             documentId
         }));
     },
     documentGeneralSettings () {
-        const DocumentSettingsGeneralView = require('../views/documentSettingsGeneral');
         this.view.renderView(new DocumentSettingsGeneralView());
     },
     documentAccountSettings () {
-        const DocumentSettingsAccountView = require('../views/documentSettingsAccount');
         this.view.renderView(new DocumentSettingsAccountView());
     },
     documentAccountDetailSettings (documentId, accountId) {
-        const DocumentSettingsAccountDetailsView = require('../views/documentSettingsAccountDetails');
         this.view.renderView(new DocumentSettingsAccountDetailsView({
             documentId,
             accountId
         }));
     },
     documentAccountSettingsAdd () {
-        const DocumentSettingsAccountAddView = require('../views/documentSettingsAccountAdd');
         this.view.renderView(new DocumentSettingsAccountAddView());
     },
     documentAccountSettingsAddCash () {
-        const DocumentSettingsAccountAddManualView = require('../views/documentSettingsAccountAddManual');
         this.view.renderView(new DocumentSettingsAccountAddManualView({
             type: 'cash'
         }));
     },
     documentAccountSettingsAddOther () {
-        const DocumentSettingsAccountAddView = require('../views/documentSettingsAccountAddManual');
         this.view.renderView(new DocumentSettingsAccountAddView({
             type: 'other'
         }));
     },
     documentAccountSettingsAddCurrent (documentId, q) {
-        const DocumentSettingsPluginAddView = require('../views/documentSettingsPluginAdd');
         this.view.renderView(new DocumentSettingsPluginAddView({type: 'account', q}));
     },
 
     documentPluginSettings (documentId) {
-        const DocumentSettingsPluginView = require('../views/documentSettingsPlugin');
         this.view.renderView(new DocumentSettingsPluginView({
             documentId
         }));
     },
     documentPluginDetailSettings (documentId, pluginId) {
-        const DocumentSettingsPluginDetailsView = require('../views/documentSettingsPluginDetails');
         this.view.renderView(new DocumentSettingsPluginDetailsView({
             documentId,
             pluginId
         }));
     },
     documentPluginSettingsAdd (documentId, q) {
-        const DocumentSettingsPluginAddView = require('../views/documentSettingsPluginAdd');
         this.view.renderView(new DocumentSettingsPluginAddView({q}));
     },
     documentPluginSettingsAddDetails (documentId, pluginId) {
-        const PluginModel = require('../models/plugin');
-        const DocumentSettingsPluginAddDetailsView = require('../views/documentSettingsPluginAddDetails');
         this.view.renderView(new DocumentSettingsPluginAddDetailsView({
             model: new PluginModel({id: pluginId})
         }));
     },
     documentPluginSettingsAddConfig (documentId, pluginId) {
-        const PluginInstanceModel = require('../models/pluginInstance');
-        const DocumentSettingsPluginAddSetupView = require('../views/documentSettingsPluginAddSetup');
         this.view.renderView(new DocumentSettingsPluginAddSetupView({
             model: new PluginInstanceModel({id: pluginId})
         }));
     },
 
     documentPermissionSettings (documentId) {
-        const DocumentSettingsPermissionsView = require('../views/documentSettingsPermissions');
         this.view.renderView(new DocumentSettingsPermissionsView({
             documentId
         }));
     },
     documentNotificationSettings (documentId) {
-        const PlaceHolderView = require('../views/placeholder');
         this.view.renderView(new PlaceHolderView({
             icon: 'cog',
             text: `Document ${documentId.substr(0, 8)}: NotificationSettings`
@@ -221,35 +220,30 @@ module.exports = Backbone.Router.extend({
         this.navigate('admin/', {trigger: true});
     },
     adminDashboard () {
-        const PlaceHolderView = require('../views/placeholder');
         this.view.renderView(new PlaceHolderView({
             icon: 'database',
             text: 'Admin Dashboard'
         }));
     },
     adminUpdates () {
-        const PlaceHolderView = require('../views/placeholder');
         this.view.renderView(new PlaceHolderView({
             icon: 'database',
             text: 'Admin Updates'
         }));
     },
     adminUsers () {
-        const PlaceHolderView = require('../views/placeholder');
         this.view.renderView(new PlaceHolderView({
             icon: 'database',
             text: 'Admin Users'
         }));
     },
     adminPlugins () {
-        const PlaceHolderView = require('../views/placeholder');
         this.view.renderView(new PlaceHolderView({
             icon: 'database',
             text: 'Admin Plugins'
         }));
     },
     adminBackups () {
-        const PlaceHolderView = require('../views/placeholder');
         this.view.renderView(new PlaceHolderView({
             icon: 'database',
             text: 'Admin Backups'
@@ -257,14 +251,12 @@ module.exports = Backbone.Router.extend({
     },
 
     accountDevices () {
-        const PlaceHolderView = require('../views/placeholder');
         this.view.renderView(new PlaceHolderView({
             icon: 'cog',
             text: 'Account Devices'
         }));
     },
     accountAuthentication () {
-        const PlaceHolderView = require('../views/placeholder');
         this.view.renderView(new PlaceHolderView({
             icon: 'cog',
             text: 'Account Authentication'
@@ -278,8 +270,9 @@ module.exports = Backbone.Router.extend({
                 location.reload();
             })
             .catch(error => {
-                const ErrorView = require('../views/error');
                 new ErrorView({error}).appendTo(this.view);
             });
     }
 });
+
+export default AppRouter;
