@@ -1,6 +1,12 @@
-module.exports = [
-    require('./budgeted-last-month'),
-    require('./spent-last-month'),
-    require('./budgeted-average'),
-    require('./spent-average')
+import BudgetedLastMonthTip from './budgeted-last-month';
+import SpentLastMonthTip from './spent-last-month';
+import BudgetedAverageTip from './budgeted-average';
+import SpentAverageTip from './spent-average';
+
+/** @type Array<BaseTip> */
+export default [
+    BudgetedLastMonthTip,
+    SpentLastMonthTip,
+    BudgetedAverageTip,
+    SpentAverageTip
 ];
