@@ -1,19 +1,19 @@
 'use strict';
 
-const View = require('./_');
+import BaseView from './_';
 
-const AppHelper = require('../helpers/app');
-const DataHelper = require('../helpers/data');
-const TemplateHelper = require('../helpers/template');
-const LoginTemplate = require('../../templates/login.html');
-const ConfigurationHelper = require('../helpers/configuration');
+import AppHelper from '../helpers/app';
+import DataHelper from '../helpers/data';
+import TemplateHelper from '../helpers/template';
+import LoginTemplate from '../../templates/login.html';
+import ConfigurationHelper from '../helpers/configuration';
 
 /**
  * @module views/login
  * @class LoginView
- * @augments View
+ * @augments BaseView
  */
-const LoginView = View.extend({
+const LoginView = BaseView.extend({
     className: 'login',
 
     render () {
@@ -32,7 +32,8 @@ const LoginView = View.extend({
                 rendered: false,
                 visible: false,
                 checked: false,
-                content: ''
+                content: '',
+                version: null
             }
         };
 
@@ -132,16 +133,17 @@ const LoginView = View.extend({
         DataHelper.login({
             email: this.data.form.username,
             password: this.data.form.password,
-            acceptedTerms: this.data.terms.checked
+            acceptedTerms: this.data.terms.checked ? this.data.terms.version : null
         }).catch((error) => {
             if(error && error.attributes && error.attributes.acceptedTerms) {
                 this.data.terms.visible = true;
                 this.data.terms.content = ConfigurationHelper.getString('login.terms.content', {
-                    tos: error.extra.tos,
-                    privacy: error.extra.privacy
+                    tos: error.extra.tos.defaultUrl,
+                    privacy: error.extra.privacy.defaultUrl
                 });
 
                 this.updateDisabledState();
+                this.data.terms.version = error.extra.version;
                 return;
             }
 
@@ -163,4 +165,4 @@ const LoginView = View.extend({
     }
 });
 
-module.exports = LoginView;
+export default LoginView;
