@@ -6,10 +6,10 @@
  */
 
 
-require('@babel/register');
-require('@babel/polyfill');
+import '@babel/register';
+import '@babel/polyfill';
 
-const DataWorker = require('./helpers/data/worker');
-const config = require('worker-config');
+import DataWorker from './helpers/data/worker';
+import config from 'worker-config';
 
 new DataWorker(self, config);
