@@ -4,10 +4,12 @@
  *
  */
 
-module.exports = {
+const config = {
     static: __dirname + '/../dest',
     languages: [
         'de-DE',
         'en-US'
     ]
 };
+
+module.exports = config;
